+++ crates/polls/tests/integration.rs
@@ -2,23 +2,28 @@
 //
 // SPDX-License-Identifier: EUPL-1.2
 
-use std::time::Duration;
+use std::{collections::HashSet, time::Duration};
 
 use opentalk_polls::*;
 use opentalk_signaling_core::module_tester::{ModuleTester, WsMessageOutgoing};
 use opentalk_test_util::*;
 use opentalk_types::signaling::polls::{
-    command::{PollsCommand, Start, Vote},
+    command::{Choices, PollsCommand, Start, Vote},
     event::{Error, PollsEvent, Started},
     Choice, ChoiceId, Item, Results,
 };
 use pretty_assertions::assert_eq;
 use serial_test::serial;
 
-async fn start_poll(module_tester: &mut ModuleTester<Polls>, live_poll: bool) -> Started {
+async fn start_poll(
+    module_tester: &mut ModuleTester<Polls>,
+    live: bool,
+    multiple_choice: bool,
+) -> Started {
     let start = PollsCommand::Start(Start {
         topic: "polling".into(),
-        live: live_poll,
+        live,
+        multiple_choice,
         choices: vec!["yes".into(), "no".into(), "maybe".into()],
         duration: Duration::from_secs(2),
     });
@@ -42,13 +47,15 @@ async fn start_poll(module_tester: &mut ModuleTester<Polls>, live_poll: bool) ->
     if let WsMessageOutgoing::Module(PollsEvent::Started(Started {
         id,
         topic,
-        live,
+        live: live_from_event,
+        multiple_choice: multiple_choice_from_event,
         choices,
         duration,
     })) = started1
     {
         assert_eq!(topic, "polling");
-        assert_eq!(live, live_poll);
+        assert_eq!(live_from_event, live);
+        assert_eq!(multiple_choice_from_event, multiple_choice);
         assert_eq!(
             choices,
             &[
@@ -72,6 +79,7 @@ async fn start_poll(module_tester: &mut ModuleTester<Polls>, live_poll: bool) ->
             id,
             topic,
             live,
+            multiple_choice,
             choices,
             duration,
         }
@@ -87,7 +95,7 @@ async fn full_poll_with_2sec_duration() {
 
     let (mut module_tester, _user1, _user2) = common::setup_users::<Polls>(&test_ctx, ()).await;
 
-    let started = start_poll(&mut module_tester, true).await;
+    let started = start_poll(&mut module_tester, true, true).await;
 
     // User 1 vote: "yes"
 
@@ -96,7 +104,9 @@ async fn full_poll_with_2sec_duration() {
             &USER_1.participant_id,
             PollsCommand::Vote(Vote {
                 poll_id: started.id,
-                choice_id: Some(ChoiceId::from(0)),
+                choices: Choices::Multiple {
+                    choice_ids: HashSet::from_iter(vec![ChoiceId::from(0)]),
+                },
             }),
         )
         .unwrap();
@@ -136,14 +146,16 @@ async fn full_poll_with_2sec_duration() {
         panic!("unexpected {update1:?}")
     }
 
-    // User 2 vote: "no"
+    // User 2 vote: "no" (via choice_id field)
 
     module_tester
         .send_ws_message(
             &USER_2.participant_id,
             PollsCommand::Vote(Vote {
                 poll_id: started.id,
-                choice_id: Some(ChoiceId::from(1)),
+                choices: Choices::Single {
+                    choice_id: ChoiceId::from(1),
+                },
             }),
         )
         .unwrap();
@@ -190,7 +202,9 @@ async fn full_poll_with_2sec_duration() {
             &USER_2.participant_id,
             PollsCommand::Vote(Vote {
                 poll_id: started.id,
-                choice_id: Some(ChoiceId::from(2)),
+                choices: Choices::Multiple {
+                    choice_ids: HashSet::from_iter(vec![ChoiceId::from(2)]),
+                },
             }),
         )
         .unwrap();
@@ -230,6 +244,55 @@ async fn full_poll_with_2sec_duration() {
         panic!("unexpected {update1:?}")
     }
 
+    // User 2 vote again: "yes" and "no"
+
+    module_tester
+        .send_ws_message(
+            &USER_2.participant_id,
+            PollsCommand::Vote(Vote {
+                poll_id: started.id,
+                choices: Choices::Multiple {
+                    choice_ids: HashSet::from_iter(vec![ChoiceId::from(0), ChoiceId::from(1)]),
+                },
+            }),
+        )
+        .unwrap();
+
+    let update1 = module_tester
+        .receive_ws_message(&USER_1.participant_id)
+        .await
+        .unwrap();
+
+    let update2 = module_tester
+        .receive_ws_message(&USER_2.participant_id)
+        .await
+        .unwrap();
+
+    assert_eq!(update1, update2);
+
+    if let WsMessageOutgoing::Module(PollsEvent::LiveUpdate(Results { id, results })) = update1 {
+        assert_eq!(id, started.id);
+        assert_eq!(
+            results,
+            &[
+                Item {
+                    id: ChoiceId::from(0),
+                    count: 2,
+                },
+                Item {
+                    id: ChoiceId::from(1),
+                    count: 1,
+                },
+                Item {
+                    id: ChoiceId::from(2),
+                    count: 0,
+                }
+            ]
+        );
+    } else {
+        panic!("unexpected {update1:?}")
+    }
+
     // User 2 vote again: invalid choice -> fails
 
     module_tester
@@ -237,7 +300,9 @@ async fn full_poll_with_2sec_duration() {
             &USER_2.participant_id,
             PollsCommand::Vote(Vote {
                 poll_id: started.id,
-                choice_id: Some(ChoiceId::from(3)),
+                choices: Choices::Multiple {
+                    choice_ids: HashSet::from_iter(vec![ChoiceId::from(3)]),
+                },
             }),
         )
         .unwrap();
@@ -260,7 +325,9 @@ async fn full_poll_with_2sec_duration() {
             &USER_2.participant_id,
             PollsCommand::Vote(Vote {
                 poll_id: started.id,
-                choice_id: None,
+                choices: Choices::Multiple {
+                    choice_ids: HashSet::new(),
+                },
             }),
         )
         .unwrap();
@@ -326,6 +393,178 @@ async fn full_poll_with_2sec_duration() {
                 Item {
                     id: ChoiceId::from(1),
                     count: 0,
+                },
+                Item {
+                    id: ChoiceId::from(2),
+                    count: 0,
+                }
+            ]
+        );
+    } else {
+        panic!("unexpected {done1:?}")
+    }
+
+    module_tester.shutdown().await.unwrap()
+}
+
+#[actix_rt::test]
+#[serial]
+async fn poll_with_invalid_multiple_choice() {
+    let test_ctx = TestContext::new().await;
+
+    let (mut module_tester, _user1, _user2) = common::setup_users::<Polls>(&test_ctx, ()).await;
+
+    let started = start_poll(&mut module_tester, true, false).await;
+
+    // User 1 vote: "yes"
+
+    module_tester
+        .send_ws_message(
+            &USER_1.participant_id,
+            PollsCommand::Vote(Vote {
+                poll_id: started.id,
+                choices: Choices::Multiple {
+                    choice_ids: HashSet::from_iter(vec![ChoiceId::from(0)]),
+                },
+            }),
+        )
+        .unwrap();
+
+    let update1 = module_tester
+        .receive_ws_message(&USER_1.participant_id)
+        .await
+        .unwrap();
+
+    let update2 = module_tester
+        .receive_ws_message(&USER_2.participant_id)
+        .await
+        .unwrap();
+
+    assert_eq!(update1, update2);
+
+    if let WsMessageOutgoing::Module(PollsEvent::LiveUpdate(Results { id, results })) = update1 {
+        assert_eq!(id, started.id);
+        assert_eq!(
+            results,
+            &[
+                Item {
+                    id: ChoiceId::from(0),
+                    count: 1,
+                },
+                Item {
+                    id: ChoiceId::from(1),
+                    count: 0,
+                },
+                Item {
+                    id: ChoiceId::from(2),
+                    count: 0,
+                }
+            ]
+        );
+    } else {
+        panic!("unexpected {update1:?}")
+    }
+
+    // User 1 vote: "yes" and "no" (invalid choice) -> fails
+
+    module_tester
+        .send_ws_message(
+            &USER_1.participant_id,
+            PollsCommand::Vote(Vote {
+                poll_id: started.id,
+                choices: Choices::Multiple {
+                    choice_ids: HashSet::from_iter(vec![ChoiceId::from(0), ChoiceId::from(1)]),
+                },
+            }),
+        )
+        .unwrap();
+
+    let error = module_tester
+        .receive_ws_message(&USER_1.participant_id)
+        .await
+        .unwrap();
+
+    if let WsMessageOutgoing::Module(PollsEvent::Error(Error::MultipleChoicesNotAllowed)) = error {
+        // OK
+    } else {
+        panic!("unexpected {error:?}")
+    }
+
+    // User 1 vote: "no"
+
+    module_tester
+        .send_ws_message(
+            &USER_1.participant_id,
+            PollsCommand::Vote(Vote {
+                poll_id: started.id,
+                choices: Choices::Multiple {
+                    choice_ids: HashSet::from_iter(vec![ChoiceId::from(1)]),
+                },
+            }),
+        )
+        .unwrap();
+
+    let update1 = module_tester
+        .receive_ws_message(&USER_1.participant_id)
+        .await
+        .unwrap();
+
+    let update2 = module_tester
+        .receive_ws_message(&USER_2.participant_id)
+        .await
+        .unwrap();
+
+    assert_eq!(update1, update2);
+
+    if let WsMessageOutgoing::Module(PollsEvent::LiveUpdate(Results { id, results })) = update1 {
+        assert_eq!(id, started.id);
+        assert_eq!(
+            results,
+            &[
+                Item {
+                    id: ChoiceId::from(0),
+                    count: 0,
+                },
+                Item {
+                    id: ChoiceId::from(1),
+                    count: 1,
+                },
+                Item {
+                    id: ChoiceId::from(2),
+                    count: 0,
+                }
+            ]
+        );
+    } else {
+        panic!("unexpected {update1:?}")
+    }
+
+    // Poll expired, getting results in `Done` event
+
+    let done1 = module_tester
+        .receive_ws_message_override_timeout(&USER_1.participant_id, Duration::from_secs(3))
+        .await
+        .unwrap();
+
+    let done2 = module_tester
+        .receive_ws_message(&USER_2.participant_id)
+        .await
+        .unwrap();
+
+    assert_eq!(done1, done2);
+
+    if let WsMessageOutgoing::Module(PollsEvent::Done(Results { id, results })) = &done1 {
+        assert_eq!(*id, started.id);
+        assert_eq!(
+            results,
+            &[
+                Item {
+                    id: ChoiceId::from(0),
+                    count: 0,
+                },
+                Item {
+                    id: ChoiceId::from(1),
+                    count: 1,
                 },
                 Item {
                     id: ChoiceId::from(2),